Created the package controller and the class userCOntroller

+++ src/main/java/br/com/bianchi/ipa/domain/User.java
@@ -1,9 +1,8 @@
 package br.com.bianchi.ipa.domain;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,12 +11,11 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private Integer Id;
     private String nome;
 
     @Column(unique = true)
     private String  email;
     private String password;
-
 
 }
